* Refactoring of tensors made * New classes for tensors created * Homogenizers are agnostic of input and output tensors

+++ Topology Optimization/Homogenization/Sources/Tensors/Tensor2VoigtConverter/Tensor2VoigtConverter.m
@@ -17,7 +17,7 @@ classdef Tensor2VoigtConverter < handle
     methods (Access = private)
         
         function vt = getTensor(obj)
-            vt = double(obj.voigtTensor);
+            vt = obj.voigtTensor;
         end
         
     end
